Conflicts:
	CRM/Contribute/BAO/Query.php
	CRM/Core/Permission.php

+++ api/v3/Contribution.php
@@ -46,6 +46,14 @@ function civicrm_api3_contribution_create(&$params) {
   _civicrm_api3_custom_format_params($params, $values, 'Contribution');
   $params = array_merge($params, $values);
 
+  if (empty($params['id'])) {
+    $op = 'add';
+  }
+  CRM_Financial_BAO_FinancialType::getAvailableFinancialTypes($types, $op);
+  if (!in_array($params['financial_type_id'], array_keys($types))) {
+    return civicrm_api3_create_error('You do not have permission to create this contribution');
+  }
+
   if (!empty($params['id']) && !empty($params['contribution_status_id'])) {
     $error = array();
     //throw error for invalid status change such as setting completed back to pending
@@ -183,6 +191,14 @@ function _civicrm_api3_contribution_create_legacy_support_45(&$params) {
 function civicrm_api3_contribution_delete($params) {
 
   $contributionID = !empty($params['contribution_id']) ? $params['contribution_id'] : $params['id'];
+  // First check contribution financial type
+  $financialType = CRM_Core_DAO::getFieldValue('CRM_Contribute_DAO_Contribution', $contributionID, 'financial_type_id');
+  
+  // Now check permissioned lineitems & permissioned contribution
+  if (!CRM_Core_Permission::check('delete contributions of type ' . CRM_Contribute_PseudoConstant::financialType($financialType)) || 
+    !CRM_Financial_BAO_FinancialType::checkPermissionedLineItems($contributionID, 'delete', FALSE)) {
+    return civicrm_api3_create_error('You do not have permission to delete this contribution');
+  }
   if (CRM_Contribute_BAO_Contribution::deleteContribution($contributionID)) {
     return civicrm_api3_create_success(array($contributionID => 1));
   }
